backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
psycopg2-binary==2.9.10
httpx==0.27.0
uvicorn[standard]==0.29.0
whitenoise==6.6.0
celery==5.3.6
redis==5.0.4
django-celery-beat==2.6.0
openai==1.25.0
tiktoken==0.8.0
pytest==8.1.2
pytest-django==4.8.0
pytest-asyncio==0.23.6
model-bakery==1.18.0
